move macro to own file
